Create ELM327_bluetooth_program_version_1
This is an AT command sending tool. First it asks for the correct serial COM port name and sets the port up then - after connecting to it - asks for the AT command. You type it in (if You forget to add CR at the end the program will do it for You) then sends it to ELM327. It waits for the response and write it for You on the Console. If You want to finish, type 'exit', it will close the port (to free up resources).